+++ SplunkAgent/Sources/SplunkAgent/Public API/SplunkRum.swift
@@ -51,6 +51,7 @@ public class SplunkRum: ObservableObject {
     var moduleConfigurations: [Any]?
     var screenNameChangeCallback: ((String) -> Void)?
     var sessionReplayDecisionMade = false
+    var userActivityTask: Task<Void, Never>?
 
     // MARK: - Internal (Modules Proxy)
 
@@ -339,4 +340,8 @@ public class SplunkRum: ObservableObject {
 
     /// A version of this agent.
     public static let version = "2.3.1"
+
+    deinit {
+        userActivityTask?.cancel()
+    }
 }